+++ packages/antwerp-ui/react-components/src/lib/atoms/checkbox/Checkbox.tsx
@@ -1,6 +1,7 @@
 import { ChangeEvent } from 'react';
 import { SCREEN_READER_CLASS } from '../../../constants/settings';
 import { classNames } from '../../../utils/dom.utils';
+import { renderRequiredAsterisk } from '../../../utils/render.utils';
 import { CheckboxProps } from './Checkbox.types';
 
 export function Checkbox({
@@ -12,7 +13,8 @@ export function Checkbox({
   name,
   disabled,
   indeterminate,
-  onChange
+  onChange,
+  required = false
 }: CheckboxProps) {
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     return onChange && onChange(event);
@@ -40,9 +42,11 @@ export function Checkbox({
         ref={setIndeterminate}
         disabled={disabled}
         onChange={handleChange}
+        required={required}
       />
       <label htmlFor={id}>
         <span className={showLabel ? undefined : SCREEN_READER_CLASS}>{label}</span>
+        {required && renderRequiredAsterisk()}
       </label>
     </div>
   );