+++ packages/antwerp-ui/react-components/src/lib/atoms/radio/RadioGroup.tsx
@@ -1,8 +1,10 @@
 import { RadioGroupProps } from './Radio.types';
 import { ChangeEvent, cloneElement, Children, ReactElement } from 'react';
 import { classNames } from '../../../utils/dom.utils';
+import { renderDescription } from '../input/input.renders';
+import { renderRequiredAsterisk } from '../../../utils/render.utils';
 
-export function RadioGroup({ label, name, children = [], value, qa, onChange }: RadioGroupProps) {
+export function RadioGroup({ label, name, children = [], value, qa, onChange, description, state, required = false }: RadioGroupProps) {
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     return onChange && onChange(event.target?.value);
   };
@@ -13,11 +15,13 @@ export function RadioGroup({ label, name, children = [], value, qa, onChange }: 
 
   return (
     <fieldset className={classes} data-qa={qa}>
-      <legend className="a-input__label">{label}</legend>
+      <legend className="a-input__label">{label}{required && renderRequiredAsterisk()}</legend>
+      {renderDescription({ description, state })}
       {Children.map(children, (child: ReactElement) => {
         return cloneElement(child, {
           name: name,
           onChange: handleChange,
+          required: required,
           ...(value ? { checked: child.props.value === value } : {}),
           ...child.props
         });