+++ packages/antwerp-ui/react-components/src/lib/molecules/autocomplete/Autocomplete.tsx
@@ -21,7 +21,8 @@ export function Autocomplete({
   qa,
   description,
   state,
-  addOn
+  addOn,
+  disabled = false
 }: AutocompleteProps) {
   const [isOpen, setIsOpen] = useState(false);
   const [results, setResults] = useState(items);
@@ -113,6 +114,7 @@ export function Autocomplete({
   };
 
   const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    if (!isOpen) return;
     const totalResults = results?.length || 0;
     switch (e.key) {
       case 'ArrowDown':
@@ -124,6 +126,7 @@ export function Autocomplete({
         setCursor(cursor === -1 ? totalResults - 1 : cursor - 1 < 0 ? totalResults - 1 : cursor - 1);
         return setIsOpen(true);
       case 'Enter':
+        e.preventDefault();
         return onEnter();
       case 'Escape':
         return closeFlyout();
@@ -160,6 +163,7 @@ export function Autocomplete({
           required={required}
           state={state}
           addOn={addOn ? { ...addOn, content: addOn.content } : undefined}
+          disabled={disabled}
         />
       }
       ref={flyoutRef}